requirements.txt
flask
skpy
python-dotenv
pymongo